[MOVE] clients,server 파일 위치 이동 및 src/api/폴더 제거
