+++ src/lib/apis/user/onboarding.ts
@@ -1,4 +1,4 @@
-import { api } from "@/api/client";
+import { api } from "@/lib/apis/client";
 import type {
   OnboardingRequest,
   OnboardingResponse,
